remove util package
