+++ go.mod
@@ -2,19 +2,17 @@ module github.com/subspacecommunity/subspace
 
 go 1.14
 
-replace github.com/subspacecommunity/subspace/util => ./util
-
 require (
 	github.com/crewjam/saml v0.3.0
 	github.com/dustin/go-humanize v1.0.0
 	github.com/gorilla/securecookie v1.1.1
 	github.com/jteeuwen/go-bindata v3.0.8-0.20180305030458-6025e8de665b+incompatible
 	github.com/julienschmidt/httprouter v1.3.0
+	github.com/kevinburke/go-bindata v3.21.0+incompatible // indirect
 	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
 	github.com/pquerna/otp v1.2.0 // indirect
 	github.com/sirupsen/logrus v1.6.0
 	github.com/skip2/go-qrcode v0.0.0-20200519171959-a3b48390827e
-	github.com/subspacecommunity/subspace/util v0.0.0-00010101000000-000000000000
 	golang.org/x/crypto v0.0.0-20200510223506-06a226fb4e37
 	golang.org/x/net v0.0.0-20200519113804-d87ec0cfa476
 	gopkg.in/alexcesaro/quotedprintable.v3 v3.0.0-20150716171945-2caba252f4dc // indirect
