--- util/go.mod
@@ -1,3 +1,0 @@
-module github.com/subspacecloud/subspace/util
-
-go 1.14
